sdfsdafihusdf

+++ resources/views/livewire/main/supplier/supplier-index.blade.php
@@ -41,8 +41,13 @@
                             <tr>
                                 <td>{{ $origin->name }} ({{ $origin->is_local ? "Local" : "Import" }})</td>
                                 <td>
-                                    <button class="btn btn-outline-danger" wire:click="deleteOrigin({{$origin->id}})" wire:confirm="Are you sure you want to delete this origin?">Delete</button>
-                                    <a class="btn btn-outline-primary" href="{{ route("main.origins.edit" , $origin->id) }}">Edit</a>
+                                    @if ($origin->restore_date == null && $origin->deleted_at == null)
+                                        <button class="btn btn-outline-danger" wire:click="deleteOrigin({{$origin->id}})" wire:confirm="Are you sure you want to delete this origin?">Delete</button>
+                                        <a class="btn btn-outline-primary" href="{{ route("main.origins.edit" , $origin->id) }}">Edit</a>
+                                    @endif
+                                    @if ($origin->restore_date  && $origin->restore_date > today())
+                                        <button class="btn btn-outline-danger" wire:click="restoreOrigin({{$origin->id}})" wire:confirm="Are you sure you want to restore this origin?">Restore ({{ number_format(now()->diffInDays($origin->restore_date), 0) }} days left)</button>
+                                    @endif
                                 </td>
                             </tr>
                         @empty
@@ -74,8 +79,13 @@
                                 <td>{{ $supplier->contact_number }}</td>
                                 <td>{{ $supplier->email }}</td>
                                 <td>
-                                    <button class="btn btn-outline-danger" wire:click="deleteSupplier({{$supplier}})" wire:confirm="Are you sure you want to delete this post?">Delete</button>
-                                    <button class="btn btn-outline-primary" wire:click="editSupplier({{$supplier->id}})">Edit</button>
+                                    @if ($supplier->deleted_at == null && $supplier->restore_date == null)
+                                        <button class="btn btn-outline-danger" wire:click="deleteSupplier({{$supplier->id}})" wire:confirm="Are you sure you want to delete this supplier?">Delete</button>
+                                        <button class="btn btn-outline-primary" wire:click="editSupplier({{$supplier->id}})">Edit</button>
+                                    @endif
+                                    @if ($supplier->deleted_at && $supplier->restore_date  && $supplier->restore_date > today())
+                                        <button class="btn btn-outline-danger" wire:click="restoreSupplier({{$supplier->id}})" wire:confirm="Are you sure you want to restore this supplier?">Restore ({{ number_format(now()->diffInDays($supplier->restore_date), 0) }} day/s left)</button>
+                                    @endif
                                 </td>
                             </tr>
                         @empty
